@@ -1,5 +1,5 @@
 import React from 'react';
-import { ListItem, ListItemText } from '@material-ui/core';
+import { ListItem, ListItemText } from '@mui/material';
 import { Link } from "react-router-dom";
 
 export default class CalcTheme extends React.Component {
